requirements.txt
ultralytics
opencv-python-headless
numpy
pillow
pandas
scikit-learn